+++ docs/CHANGELOG.md
@@ -19,11 +19,13 @@ Changes are grouped as follows:
 ### Added
 - All `neat.read` methods now supports reading from a URL.
 - `neat.prepare.instances.connection_to_data_type(...)` added to convert connection to data type.
+- `neat.prepare.instances.classic_to_core()` which bundles the preparation from classic to core.
 
 ### Fixed
 - `neat.read.yaml(...)` no longer raises a `ValueError` if a view is referencing an enum property in a container
   that is not part of the model.
 - `neat.read.yaml('....zip', format='toolkit')` now correctly puts containers into a subfolder in the zip file.
+- `neat.to.cdf.data_model(...)` now correctly handles list of json objects in the `properties` field of a view.
 
 ### Improved
 - `neat.read.yaml(..., format="toolkit")` no longer removes direct relations sources that points to views that are not